Updated exploit.html
Added function to handle better 'cmd' variable. The first check to see if the cmd length is an even or odd number, and if it’s odd, will add a single space character to the end of the command to convert it to even length. This is used to fill in the last buffer space as it works in 4 bytes at a time. From there, it will do a series of loops for the conversion into the format it needs to run.